refactoring architecture, categorising methods into classes and then into packages
