+++ src/main/java/com/mindlinksoft/recruitment/mychat/CLI.java
@@ -1,0 +1,9 @@
+package com.mindlinksoft.recruitment.mychat;
+
+public class CLI {
+    public static void main(String[] args) throws Exception {
+        String[] arguments = new String[]{"D:\\Desktop\\my-chat\\chat.txt", "D:\\Desktop\\my-chat\\chat.json"};
+        ConversationExporter.main(arguments);
+    }
+}
+
