+++ src/main/java/com/mindlinksoft/recruitment/mychat/ConversationExporter.java
@@ -1,106 +1,28 @@
 package com.mindlinksoft.recruitment.mychat;
 
-import java.io.*;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
 
-/**
- * Represents a conversation exporter that can read a conversation and write it out in JSON.
- */
+import com.mindlinksoft.recruitment.mychat.Objects.Conversation;
+import com.mindlinksoft.recruitment.mychat.Objects.ConversationExporterConfiguration;
+import com.mindlinksoft.recruitment.mychat.Utilities.ReadWrite;
+import com.mindlinksoft.recruitment.mychat.Utilities.CommandLineArgumentParser;
+
 public class ConversationExporter {
+    private ReadWrite readWrite = new ReadWrite();
 
-    /**
-     * The application entry point.
-     *
-     * @param args The command line arguments.
-     * @throws Exception Thrown when something bad happens.
-     */
     public static void main(String[] args) throws Exception {
+
         ConversationExporter exporter = new ConversationExporter();
         ConversationExporterConfiguration configuration = new CommandLineArgumentParser().parseCommandLineArguments(args);
 
         exporter.exportConversation(configuration.inputFilePath, configuration.outputFilePath);
     }
 
-    /**
-     * Exports the conversation at {@code inputFilePath} as JSON to {@code outputFilePath}.
-     *
-     * @param inputFilePath  The input file path.
-     * @param outputFilePath The output file path.
-     * @throws Exception Thrown when something bad happens.
-     */
-    public void exportConversation(String inputFilePath, String outputFilePath) throws Exception {
-        Conversation conversation = this.readConversation(inputFilePath);
+    void exportConversation(String inputFilePath, String outputFilePath) throws Exception {
+        Conversation conversation = this.readWrite.readConversation(inputFilePath);
 
-        this.writeConversation(conversation, outputFilePath);
+        readWrite.writeConversation(conversation, outputFilePath);
 
         // TODO: More logging
         System.out.println("Conversation exported from '" + inputFilePath + "' to '" + outputFilePath);
-    }
-
-    /**
-     * Helper method to write the given {@code conversation} as JSON to the given {@code outputFilePath}.
-     *
-     * @param conversation   The conversation to write.
-     * @param outputFilePath The file path where the conversation should be written.
-     * @throws Exception Thrown when something bad happens.
-     */
-    public void writeConversation(Conversation conversation, String outputFilePath) throws Exception {
-        // TODO: Do we need both to be resources, or will buffered writer close the stream?
-        try (OutputStream os = new FileOutputStream(outputFilePath, true);
-             BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(os))) {
-
-            // TODO: Maybe reuse this? Make it more testable... // DONE
-            BuildCreateGson gson = new BuildCreateGson();
-            bw.write(gson.g.toJson(conversation));
-
-        } catch (FileNotFoundException e) {
-            // TODO: Maybe include more information? // DONE
-            throw new IllegalArgumentException("Output file path argument '" + outputFilePath + "' was not found.");
-        } catch (IOException e) {
-            // TODO: Should probably throw different exception to be more meaningful :/ // DONE
-            throw new IOException("Could not write this file.");
-        }
-    }
-
-    /**
-     * Represents a helper to read a conversation from the given {@code inputFilePath}.
-     *
-     * @param inputFilePath The path to the input file.
-     * @return The {@link Conversation} representing by the input file.
-     * @throws Exception Thrown when something bad happens.
-     */
-    public Conversation readConversation(String inputFilePath) throws Exception {
-        try (InputStream is = new FileInputStream(inputFilePath);
-             BufferedReader r = new BufferedReader(new InputStreamReader(is))) {
-
-            List<Message> messages = new ArrayList<Message>();
-
-            String conversationName = r.readLine();
-            String line;
-
-            while ((line = r.readLine()) != null) {
-                String[] split = line.split(" ");
-
-                StringBuilder content = new StringBuilder();
-
-                for (int i = 0; i < split.length; i++) {
-                    if (i == 2) {
-                        content.append(split[i]);
-                    }
-                    if (i > 2) {
-                        content.append(" ").append(split[i]);
-                    }
-                }
-                messages.add(new Message(Instant.ofEpochSecond(Long.parseUnsignedLong(split[0])), split[1], content.toString()));
-            }
-
-            return new Conversation(conversationName, messages);
-        } catch (FileNotFoundException e) {
-            throw new IllegalArgumentException("Input file path argument '" + inputFilePath + "' was not found.");
-        } catch (IOException e) {
-            throw new IOException("Could not read this file.");
-        }
     }
 }
