+++ src/main/java/com/mindlinksoft/recruitment/mychat/Objects/Conversation.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.Objects;
 
 import java.util.Collection;
