+++ src/main/java/com/mindlinksoft/recruitment/mychat/Objects/Message.java
@@ -1,4 +1,4 @@
-package com.mindlinksoft.recruitment.mychat;
+package com.mindlinksoft.recruitment.mychat.Objects;
 
 import java.time.Instant;
